Компонент для кнопок и таймлайна в плеере
