+++ static/js/views/player/player.js
@@ -7,6 +7,7 @@ import PlaylistComponent from '@components/playlist_component/playlist_component
 import TopTrackComponent from '@components/top_track_component/top_track_component';
 import TrackListComponent from
     '@components/player_track_list_component/player_track_list_component';
+import PlayerControlComponent from '@components/player_control_component/player_control_component';
 
 /**
  *  вью для плеера
@@ -19,16 +20,10 @@ export default class PlayerView extends BaseView {
         super(player);
         this.eventBus = eventBus;
         this.expanded = false;
-        this.timelineDrag = false;
-        this.volumeDrag = false;
-        this.playing = false;
-        this.shuffled = false;
-        this.repeatState = 0;
-        this.muted = false;
-        this.volume = 1;
         this.locked = true;
         this.topTrackComponent = new TopTrackComponent(eventBus);
         this.trackListComponent = new TrackListComponent(eventBus);
+        this.playerControlComponent = new PlayerControlComponent(eventBus);
         this._choosePlaylist = new ChoosePlaylist(eventBus, PLAYER);
         this._playlistComponent = new PlaylistComponent(
             this.trackListComponent.setEventListeners.bind(this));
@@ -40,18 +35,7 @@ export default class PlayerView extends BaseView {
             this.trackListComponent.removeFromTracklist.bind(this));
         this.eventBus.on(PLAYER.REMOVE_FROM_TRACKLIST_ALL,
             this.trackListComponent.removeFromTracklistAll.bind(this));
-        this.eventBus.on(PLAYER.DRAW_PLAY, this.drawPlay.bind(this));
-        this.eventBus.on(PLAYER.DRAW_PAUSE, this.drawPause.bind(this));
-        this.eventBus.on(PLAYER.TRACK_UPDATE, this.updateTrack.bind(this));
         this.eventBus.on(PLAYER.MOVE_MARKER, this.moveMarker.bind(this));
-        this.eventBus.on(PLAYER.DRAW_TIMELINE, this.drawTimeline.bind(this));
-        this.eventBus.on(PLAYER.DRAW_SHUFFLE, this.drawShuffle.bind(this));
-        this.eventBus.on(PLAYER.DRAW_UNSHUFLE, this.drawUnshuffle.bind(this));
-        this.eventBus.on(PLAYER.DRAW_REPEAT, this.drawRepeat.bind(this));
-        this.eventBus.on(PLAYER.DRAW_REPEAT_ONE, this.drawRepeatOne.bind(this));
-        this.eventBus.on(PLAYER.DRAW_UNREPEAT, this.drawUnrepeat.bind(this));
-        this.eventBus.on(PLAYER.DRAW_MUTE, this.drawMute.bind(this));
-        this.eventBus.on(PLAYER.DRAW_UNMUTE, this.drawUnmute.bind(this));
         globalEventBus.on(GLOBAL.COLLAPSE, this.collapse.bind(this));
     }
 
@@ -63,6 +47,7 @@ export default class PlayerView extends BaseView {
     render(root, url) {
         super.render(document.getElementsByClassName(DOM.PLAYER)[0]);
         this.topTrackComponent.render();
+        this.playerControlComponent.render();
         this.setEventListeners();
         this.resize();
     }
@@ -80,14 +65,15 @@ export default class PlayerView extends BaseView {
         this.root.style.left = left.toString() + 'px';
         const top = NAVBAR.HEIGHT;
         const height = document.documentElement.clientHeight - top;
-        document.getElementsByTagName('audio')[0].volume = this.volume;
-        this.drawVolume(height);
+        document.getElementsByTagName('audio')[0].volume = this.playerControlComponent.volume;
+        this.playerControlComponent.drawVolume(height);
         this.root.style.top = top.toString() + 'px';
         this.root.style.height = height.toString() + 'px';
         document.getElementsByClassName('player-trigger')[0]
             .style.height = height.toString() + 'px';
-        this.drawVolume(document.getElementsByClassName('volume-scale-back')[0]
-            .getBoundingClientRect().height * this.volume);
+        this.playerControlComponent.drawVolume(document
+            .getElementsByClassName('volume-scale-back')[0]
+            .getBoundingClientRect().height * this.playerControlComponent.volume);
     }
 
     /**
@@ -97,7 +83,6 @@ export default class PlayerView extends BaseView {
         window.addEventListener('resize', this.resize.bind(this));
         document.getElementsByTagName('body')[0]
             .addEventListener('DOMSubtreeModified', this.resize.bind(this));
-        window.addEventListener('mouseup', this.windowMouseUp.bind(this));
         document.getElementsByTagName('audio')[0]
             .addEventListener('timeupdate', this.audioTimeUpdate.bind(this));
         document.getElementsByTagName('audio')[0]
@@ -114,77 +99,8 @@ export default class PlayerView extends BaseView {
             .addEventListener('mouseout', this.triggerMouseOut.bind(this));
         document.getElementsByClassName('trigger-button')[0]
             .addEventListener('click', this.triggerClick.bind(this));
-        document.getElementsByClassName('play-pause')[0]
-            .addEventListener('click', this.playPauseButtonClick.bind(this));
-        document.getElementById('prev').addEventListener('click', this.prevButtonClick.bind(this));
-        document.getElementById('next').addEventListener('click', this.nextButtonClick.bind(this));
-        document.getElementsByClassName('timeline')[0]
-            .addEventListener('mouseover', this.timelineMouseOver.bind(this));
-        document.getElementsByClassName('timeline')[0]
-            .addEventListener('mouseout', this.timelineMouseOut.bind(this));
-        document.getElementsByClassName('timeline-back')[0]
-            .addEventListener('mousedown', this.timelineMouseDown.bind(this));
-        document.getElementsByClassName('timeline-back')[0]
-            .onmouseup = (event) => this.timelineMouseUp(event);
-        document.getElementsByClassName('timeline-front')[0]
-            .addEventListener('mousedown', this.timelineMouseDown.bind(this));
-        document.getElementsByClassName('timeline-front')[0]
-            .onmouseup = (event) => this.timelineMouseUp(event);
-        document.getElementsByClassName('timeline-back')[0]
-            .onmousemove = (event) => this.timelineMouseMove(event);
-        document.getElementsByClassName('timeline-front')[0]
-            .onmousemove = (event) => this.timelineMouseMove(event);
-        document.getElementsByClassName('timeline-back')[0]
-            .onclick = (event) => this.timelineClick(event);
-        document.getElementsByClassName('timeline-front')[0]
-            .onclick = (event) => this.timelineClick(event);
-        document.getElementsByClassName('shuffle')[0]
-            .addEventListener('mouseover', this.shuffleButtonMouseOver.bind(this));
-        document.getElementsByClassName('shuffle')[0]
-            .addEventListener('mouseout', this.shuffleButtonMouseOut.bind(this));
-        document.getElementsByClassName('shuffle')[0]
-            .addEventListener('click', this.shuffleButtonClick.bind(this));
-        document.getElementsByClassName('repeat')[0]
-            .addEventListener('mouseover', this.repeatButtonMouseOver.bind(this));
-        document.getElementsByClassName('repeat')[0]
-            .addEventListener('mouseout', this.repeatButtonMouseOut.bind(this));
-        document.getElementsByClassName('repeat')[0]
-            .addEventListener('click', this.repeatButtonClick.bind(this));
-        document.getElementsByClassName('volume')[0]
-            .addEventListener('mouseover', this.volumeButtonMouseOver.bind(this));
-        document.getElementsByClassName('volume')[0]
-            .addEventListener('mouseout', this.volumeButtonMouseOut.bind(this));
-        document.getElementsByClassName('volume')[0]
-            .addEventListener('click', this.volumeButtonClick.bind(this));
-        document.getElementsByClassName('volume-scale')[0]
-            .addEventListener('mouseover', this.volumeButtonMouseOver.bind(this));
-        document.getElementsByClassName('volume-scale')[0]
-            .addEventListener('mouseout', this.volumeButtonMouseOut.bind(this));
-        document.getElementsByClassName('volume-scale-back')[0]
-            .addEventListener('mousedown', this.volumeMouseDown.bind(this));
-        document.getElementsByClassName('volume-scale-back')[0]
-            .onmouseup = (event) => this.volumeMouseUp(event);
-        document.getElementsByClassName('volume-scale-front')[0]
-            .addEventListener('mousedown', this.volumeMouseDown.bind(this));
-        document.getElementsByClassName('volume-scale-front')[0]
-            .onmouseup = (event) => this.volumeMouseUp(event);
-        document.getElementsByClassName('volume-scale-back')[0]
-            .onclick = (event) => this.volumeScaleClick(event);
-        document.getElementsByClassName('volume-scale-front')[0]
-            .onclick = (event) => this.volumeScaleClick(event);
-        document.getElementsByClassName('volume-scale-back')[0]
-            .onmousemove = (event) => this.volumeMouseMove(event);
-        document.getElementsByClassName('volume-scale-front')[0]
-            .onmousemove = (event) => this.volumeMouseMove(event);
         window.onwheel = (event) => this.trackListComponent.trackListWheel(event);
-    }
-
-    /**
-     * Слушает отпускание клавиши мыши
-     */
-    windowMouseUp() {
-        this.timelineDrag = false;
-        this.volumeDrag = false;
+        this.playerControlComponent.setEventListeners();
     }
 
     /**
@@ -198,8 +114,8 @@ export default class PlayerView extends BaseView {
         ) {
             return;
         }
-        this.drawTimeline(document.getElementsByTagName('audio')[0].currentTime /
-            document.getElementsByTagName('audio')[0].duration);
+        this.playerControlComponent.drawTimeline(document.getElementsByTagName('audio')[0]
+            .currentTime / document.getElementsByTagName('audio')[0].duration);
     }
 
     /**
@@ -256,249 +172,6 @@ export default class PlayerView extends BaseView {
     }
 
     /**
-     * Слушает клик по кнопке воспроизвдения/паузы
-     */
-    playPauseButtonClick() {
-        if (this.playing) {
-            this.eventBus.emit(PLAYER.PAUSE);
-        } else {
-            this.eventBus.emit(PLAYER.PLAY);
-        }
-    }
-
-    /**
-     * Слушает клик по кнопке включения предыдущего трека
-     */
-    prevButtonClick() {
-        this.eventBus.emit(PLAYER.PREVIOUS);
-    }
-
-    /**
-     * Слушает клик по кнопке включения следующего трека
-     */
-    nextButtonClick() {
-        this.eventBus.emit(PLAYER.NEXT, 'click');
-    }
-
-    /**
-     * Слушает вход курсора в зону таймлайна
-     */
-    timelineMouseOver() {
-        document.getElementsByClassName('current-time')[0].style.fontSize = '11px';
-        document.getElementsByClassName('duration')[0].style.fontSize = '11px';
-    }
-
-    /**
-     * Слушает выход курсора из зоны таймлайна
-     */
-    timelineMouseOut() {
-        document.getElementsByClassName('current-time')[0].style.fontSize = '0';
-        document.getElementsByClassName('duration')[0].style.fontSize = '0';
-    }
-
-    /**
-     * Слушает нажатие клавиши мыши на таймлайне
-     */
-    timelineMouseDown() {
-        this.timelineDrag = true;
-    }
-
-    /**
-     * Слушает отпускание клавиши мыши на таймлайне
-     * @param {Object} event
-     */
-    timelineMouseUp(event) {
-        this.timelineDrag = false;
-        const width = event.clientX - document.getElementsByClassName('timeline-back')[0]
-            .getBoundingClientRect().x;
-        const ratio = width / document.getElementsByClassName('timeline-back')[0]
-            .getBoundingClientRect().width;
-        this.eventBus.emit(PLAYER.REWIND, ratio);
-    }
-
-    /**
-     * Слушает движение мыши по таймлайну
-     * @param {Object} event
-     */
-    timelineMouseMove(event) {
-        if (this.timelineDrag) {
-            const width = event.clientX - document.getElementsByClassName('timeline-back')[0]
-                .getBoundingClientRect().x;
-            const ratio = width / document.getElementsByClassName('timeline-back')[0]
-                .getBoundingClientRect().width;
-            this.drawTimeline(ratio);
-        }
-    }
-
-    /**
-     * Слушает клик по таймлайну
-     * @param {Object} event
-     */
-    timelineClick(event) {
-        const width = event.clientX - document.getElementsByClassName('timeline-back')[0]
-            .getBoundingClientRect().x;
-        const ratio = width / document.getElementsByClassName('timeline-back')[0]
-            .getBoundingClientRect().width;
-        this.eventBus.emit(PLAYER.REWIND, ratio);
-    }
-
-    /**
-     * Слушает вход курсора на кнопку перемешивания
-     */
-    shuffleButtonMouseOver() {
-        if (!this.shuffled) {
-            document.getElementsByClassName('shuffle')[0].style.opacity = '1';
-        }
-    }
-
-    /**
-     * Слушает выход курсора с кнопки перемешивания
-     */
-    shuffleButtonMouseOut() {
-        if (!this.shuffled) {
-            document.getElementsByClassName('shuffle')[0].style.opacity = '0.4';
-        }
-    }
-
-    /**
-     * Слушает клик по кнопке перемешивания
-     */
-    shuffleButtonClick() {
-        if (!this.shuffled) {
-            this.eventBus.emit(PLAYER.SHUFFLE, 'first');
-        } else {
-            this.eventBus.emit(PLAYER.UNSHUFFLE);
-        }
-    }
-
-    /**
-     * Слушает вход курсора на кнопку зацикливания
-     */
-    repeatButtonMouseOver() {
-        if (this.repeatState === 0) {
-            document.getElementsByClassName('repeat')[0].style.opacity = '1';
-        }
-    }
-
-    /**
-     * Слушает выход курсора с кнопки зацикливания
-     */
-    repeatButtonMouseOut() {
-        if (this.repeatState === 0) {
-            document.getElementsByClassName('repeat')[0].style.opacity = '0.4';
-        }
-    }
-
-    /**
-     * Слушает клик по кнопке зацикливания
-     */
-    repeatButtonClick() {
-        switch (this.repeatState) {
-        case 0:
-            this.eventBus.emit(PLAYER.REPEAT);
-            break;
-        case 1:
-            this.eventBus.emit(PLAYER.REPEAT_ONE);
-            break;
-        case 2:
-            this.eventBus.emit(PLAYER.UNREPEAT);
-            break;
-        }
-    }
-
-    /**
-     * Слушает вход курсора на кнопку громкости
-     */
-    volumeButtonMouseOver() {
-        document.getElementsByClassName('volume-scale')[0]
-            .style.transitionProperty = 'opacity, top';
-        document.getElementsByClassName('volume-scale')[0].style.visibility = 'visible';
-        document.getElementsByClassName('volume-scale')[0].style.opacity = '1';
-        document.getElementsByClassName('volume-scale')[0].style.top = '-52px';
-        document.getElementsByClassName('volume')[0].style.opacity = '1';
-    }
-
-    /**
-     * Слушает выход курсора с кнопки громкости
-     */
-    volumeButtonMouseOut() {
-        document.getElementsByClassName('volume-scale')[0]
-            .style.transitionProperty = 'opacity, visibility, top';
-        document.getElementsByClassName('volume-scale')[0].style.visibility = 'hidden';
-        document.getElementsByClassName('volume-scale')[0].style.opacity = '0';
-        document.getElementsByClassName('volume-scale')[0].style.top = '-42px';
-        document.getElementsByClassName('volume')[0].style.opacity = '0.4';
-    }
-
-    /**
-     * Слушает нажатие клавиши мыши на шкале громкости
-     */
-    volumeMouseDown() {
-        this.volumeDrag = true;
-    }
-
-    /**
-     * Слушает отпускание клавиши мыши на шкале громкости
-     * @param {Object} event
-     */
-    volumeMouseUp(event) {
-        this.volumeDrag = false;
-        this.muted = false;
-        const height = document.getElementsByClassName('volume-scale-back')[0]
-            .getBoundingClientRect().height -
-            (event.clientY - document.getElementsByClassName('volume-scale-back')[0]
-                .getBoundingClientRect().y);
-        this.volume = height / document.getElementsByClassName('volume-scale-back')[0]
-            .getBoundingClientRect().height;
-        document.getElementsByTagName('audio')[0].volume = this.volume;
-        this.drawVolume(height);
-    }
-
-    /**
-     * Слушает движение мыши по шкале громкости
-     * @param {Object} event
-     */
-    volumeMouseMove(event) {
-        if (this.volumeDrag) {
-            const height = document.getElementsByClassName('volume-scale-back')[0]
-                .getBoundingClientRect().height -
-                (event.clientY - document.getElementsByClassName('volume-scale-back')[0]
-                    .getBoundingClientRect().y);
-            this.volume = height / document.getElementsByClassName('volume-scale-back')[0]
-                .getBoundingClientRect().height;
-            document.getElementsByTagName('audio')[0].volume = this.volume;
-            this.drawVolume(height);
-        }
-    }
-
-    /**
-     * Слушает клик по шкале громкости
-     * @param {Object} event
-     */
-    volumeScaleClick(event) {
-        const height = document.getElementsByClassName('volume-scale-back')[0]
-            .getBoundingClientRect().height -
-            (event.clientY - document.getElementsByClassName('volume-scale-back')[0]
-                .getBoundingClientRect().y);
-        this.volume = height / document.getElementsByClassName('volume-scale-back')[0]
-            .getBoundingClientRect().height;
-        document.getElementsByTagName('audio')[0].volume = this.volume;
-        this.drawVolume(height);
-        this.drawUnmute();
-    }
-
-    /**
-     * Слушает клик по кнопке громкости
-     */
-    volumeButtonClick() {
-        if (this.muted) {
-            this.eventBus.emit(PLAYER.UNMUTE);
-        } else {
-            this.eventBus.emit(PLAYER.MUTE);
-        }
-    }
-
-    /**
      * Получение плейлистов пользователя
      */
     getProfilePlaylists() {
@@ -521,37 +194,6 @@ export default class PlayerView extends BaseView {
         if (this.expanded) {
             this.triggerClick();
         }
-    }
-
-    /**
-     * Рисует кнопку воспроизвдения/паузы как плей для режима паузы
-     * (воспроизвдение по нажатию)
-     */
-    drawPlay() {
-        document.getElementsByClassName('play-pause')[0].src = '/static/img/play.svg';
-        this.playing = false;
-    }
-
-    /**
-     * Рисует кнопку воспроизвдения/паузы как пауза для режима плей
-     * (пауза по нажатию)
-     */
-    drawPause() {
-        document.getElementsByClassName('play-pause')[0].src = '/static/img/pause.svg';
-        this.playing = true;
-    }
-
-    /**
-     * Обновляет текущий воспроизводимый трек
-     * @param {Object} track
-     */
-    updateTrack(track) {
-        this.topTrackComponent.updateTrack(track);
-        const minutes = Math.floor(track.duration / 60);
-        const seconds = Math.floor(track.duration % 60);
-        document.getElementsByClassName('duration')[0].innerHTML = minutes.toString() + ':' +
-            (seconds < 10 ? '0' : '') + seconds.toString();
-        document.getElementsByClassName('current-time')[0].innerHTML = '0:00';
     }
 
     /**
@@ -585,100 +227,5 @@ export default class PlayerView extends BaseView {
                 track1.style.marginLeft = '0';
             }
         }, 250);
-    }
-
-    /**
-     * Рисует таймлайн в конкретном положении
-     * @param {number} ratio
-     */
-    drawTimeline(ratio) {
-        const width = ratio * document.getElementsByClassName('timeline-back')[0]
-            .getBoundingClientRect().width;
-        document.getElementsByClassName('timeline-front')[0].style.width = width.toString() + 'px';
-        const minutes = Math.floor((ratio *
-            document.getElementsByTagName('audio')[0].duration) / 60);
-        const seconds = Math.floor((ratio *
-            document.getElementsByTagName('audio')[0].duration) % 60);
-        document.getElementsByClassName('current-time')[0].innerHTML = minutes.toString() + ':' +
-            (seconds < 10 ? '0' : '') + seconds.toString();
-    }
-
-    /**
-     * Рисует кнопку перемешивания в режиме перемешивания (активной)
-     */
-    drawShuffle() {
-        document.getElementsByClassName('shuffle')[0].style.opacity = '1';
-        this.shuffled = true;
-    }
-
-    /**
-     * Рисует кнопку перемешивания в режиме воспроизвдения подряд (неактивной)
-     */
-    drawUnshuffle() {
-        document.getElementsByClassName('shuffle')[0].style.opacity = '0.4';
-        this.shuffled = false;
-    }
-
-    /**
-     * Рисует кнопку зацикливания в режиме зацикливания всего плейлиста (активной, пустой внутри)
-     */
-    drawRepeat() {
-        document.getElementsByClassName('repeat')[0].style.opacity = '1';
-        this.repeatState = 1;
-    }
-
-    /**
-     * Рисует кнопку зацикливания в режиме зацикливания одного трека (активной, с единичкой внутри)
-     */
-    drawRepeatOne() {
-        document.getElementsByClassName('repeat')[0].src = '/static/img/repeat_one.svg';
-        this.repeatState = 2;
-    }
-
-    /**
-     * Рисует кнопку зацикливания в режиме одноразового проигрывания (неактивной)
-     */
-    drawUnrepeat() {
-        document.getElementsByClassName('repeat')[0].src = '/static/img/repeat.svg';
-        document.getElementsByClassName('repeat')[0].style.opacity = '0.4';
-        this.repeatState = 0;
-    }
-
-    /**
-     * Рисует кнопку громкости в беззучном режиме (без волн)
-     */
-    drawMute() {
-        document.getElementsByClassName('volume')[0].src = '/static/img/volume_mute.svg';
-        this.drawVolume(0);
-        this.muted = true;
-    }
-
-    /**
-     * Рисует кнопку громкости в режиме со звуком (с волнами)
-     */
-    drawUnmute() {
-        if (this.volume <= 0.5) {
-            document.getElementsByClassName('volume')[0].src = '/static/img/volume_down.svg';
-        } else {
-            document.getElementsByClassName('volume')[0].src = '/static/img/volume_up.svg';
-        }
-        this.drawVolume(document.getElementsByClassName('volume-scale-back')[0]
-            .getBoundingClientRect().height * this.volume);
-        this.muted = false;
-    }
-
-    /**
-     * Рисует шкалу громкости с определённым значением
-     * @param {number} height
-     */
-    drawVolume(height) {
-        if (height === 0) {
-            document.getElementsByClassName('volume-scale-front')[0].style.height = '0';
-            return;
-        }
-        document.getElementsByClassName('volume-scale-front')[0].style.top = 0 - height.toString() +
-            'px';
-        document.getElementsByClassName('volume-scale-front')[0].style.height = height.toString() +
-            'px';
     }
 }
